Move schemas into their own subdir
