+++ src/exceptions/generate.ts
@@ -1,5 +1,5 @@
+import { annotatedHearingOutcomeSchema } from "src/schemas/annotatedHearingOutcome"
 import type { AnnotatedHearingOutcome } from "src/types/AnnotatedHearingOutcome"
-import { annotatedHearingOutcomeSchema } from "src/types/AnnotatedHearingOutcome"
 import type Exception from "src/types/Exception"
 import { ExceptionCode } from "src/types/ExceptionCode"
 import type { ZodIssue } from "zod"
