+++ src/types/Exception.ts
@@ -1,12 +1,5 @@
-import { z } from "zod"
-import { ExceptionCode } from "./ExceptionCode"
-
-const exceptionPathSchema = z.array(z.number().or(z.string()))
-
-export const exceptionSchema = z.object({
-  code: z.nativeEnum(ExceptionCode),
-  path: exceptionPathSchema
-})
+import type { exceptionPathSchema, exceptionSchema } from "src/schemas/exception"
+import type { z } from "zod"
 
 export type ExceptionPath = z.infer<typeof exceptionPathSchema>
 type Exception = z.infer<typeof exceptionSchema>
